+++ src/views/Grants/ProgramsAll.vue
@@ -58,6 +58,7 @@
                   <BaseCardExpandable
                     :item="program"
                     :openSearch="true"
+                    :showLink="false"
                   ></BaseCardExpandable></div
               ></v-col>
             </v-row>